requirements.txt
click
pymongo[srv]
discord.py==1.2.3